backend/backend/requirements.txt
numpy==1.21.6
pandas==1.3.3
scikit-learn==0.24.2
scikit-surprise==1.1.1
flask==2.0.2
werkzeug==2.0.3
Django==3.2.9